+++ DependencyInjection/Compiler/RegisterProvidersPass.php
@@ -1,0 +1,34 @@
+<?php declare(strict_types=1);
+/**
+ * @author    Igor Nikolaev <igor.sv.n@gmail.com>
+ * @copyright Copyright (c) 2020, Darvin Studio
+ * @link      https://www.darvin-studio.ru
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Darvin\MenuBundle\DependencyInjection\Compiler;
+
+use Darvin\MenuBundle\DependencyInjection\DarvinMenuExtension;
+use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Reference;
+
+/**
+ * Register menu providers compiler pass
+ */
+class RegisterProvidersPass implements CompilerPassInterface
+{
+    /**
+     * {@inheritDoc}
+     */
+    public function process(ContainerBuilder $container): void
+    {
+        $registry = $container->getDefinition('darvin_menu.provider_registry');
+
+        foreach (array_keys($container->findTaggedServiceIds(DarvinMenuExtension::TAG_PROVIDER)) as $id) {
+            $registry->addMethodCall('addProvider', [new Reference($id)]);
+        }
+    }
+}